+++ lightly_studio/src/lightly_studio/core/annotation/__init__.py
@@ -1,3 +1,4 @@
 from .annotation import Annotation
+from .annotation_label import AnnotationLabel
 
-__all__ = ["Annotation"]
+__all__ = ["Annotation", "AnnotationLabel"]